Article增加comment_count字段，ArticleComment解决循环依赖

+++ src/main/java/ski/mashiro/dto/ArticlePreviewDTO.java
@@ -15,7 +15,7 @@ public class ArticlePreviewDTO {
     private Long articleId;
     private String title;
     private String previewContent;
-    private Long commentCount;
+    private Integer commentCount;
     private List<String> tag;
     private LocalDateTime createTime;
